Rename back utopia-flags to comment-flags
