+++ editor/src/core/model/conditionals.ts
@@ -16,7 +16,10 @@ import type { ElementPathTree, ElementPathTrees } from '../shared/element-path-t
 import { getUtopiaID } from '../shared/uid-utils'
 import type { Optic } from '../shared/optics/optics'
 import { fromField, fromTypeGuard } from '../shared/optics/optic-creators'
-import { findUtopiaCommentFlag, isUtopiaPropOrCommentFlagConditional } from '../shared/utopia-flags'
+import {
+  findUtopiaCommentFlag,
+  isUtopiaPropOrCommentFlagConditional,
+} from '../shared/comment-flags'
 import { isLeft, isRight } from '../shared/either'
 import { MetadataUtils } from './element-metadata-utils'
 import { forceNotNull } from '../shared/optional-utils'
