+++ editor/src/core/shared/comment-flags.ts
